refactor: split wikr monolith and bump version to 0.8.2
