+++ io_util.go
@@ -1,0 +1,40 @@
+package main
+
+import (
+	"bufio"
+	"io"
+	"os"
+)
+
+func isTerminal(w io.Writer) bool {
+	f, ok := w.(*os.File)
+	if !ok {
+		return false
+	}
+	stat, err := f.Stat()
+	if err != nil {
+		return false
+	}
+	return (stat.Mode() & os.ModeCharDevice) != 0
+}
+
+func readerFrom(in io.Reader) *bufio.Reader {
+	if r, ok := in.(*bufio.Reader); ok {
+		return r
+	}
+	return bufio.NewReader(in)
+}
+
+func truncateWithEllipsis(s string, max int) string {
+	if max <= 0 {
+		return ""
+	}
+	runes := []rune(s)
+	if len(runes) <= max {
+		return s
+	}
+	if max <= 3 {
+		return string(runes[:max])
+	}
+	return string(runes[:max-3]) + "..."
+}
